+++ spec/spec_helper.rb
@@ -1,8 +1,19 @@
 # frozen_string_literal: true
 
-require "rails"
 require "timecop"
-require "active_support"
+
+unless ENV['REALREDIS']
+  require "fakeredis"
+  require "fakeredis/rspec"
+end
+
+begin
+  require "rails"
+  require "active_support"
+rescue LoadError
+  puts "No rails"
+end
+
 require "deprecation_collector"
 
 ENV["RAILS_ENV"] = "test"